change schema of promotion, add create,remove,filterBillByDate at controller
(*) change schema of promotion
(+) create,delete,filterByDate

+++ src/Model/Product/Warehouse.java
@@ -37,7 +37,7 @@ public class Warehouse {
 				p.getProductStockInfo().setNumStock(rs.getInt("numStock"));
 
 				p.getPromotion().setPromoID(rs.getInt("idPromo"));
-				p.getPromotion().setPromoDiscount(rs.getLong("discount"));
+				p.getPromotion().setPromoDiscount(rs.getInt("discount"));
 
 				allProducts.add(p);
 			}
